feat: onboarding sidebar
