+++ frontend/src/screens/wallet/OnboardingChecklist.tsx
@@ -1,3 +1,5 @@
+// src/components/OnboardingChecklist.tsx
+
 import { ChevronRight, Circle, CircleCheck } from "lucide-react";
 import { Link } from "react-router-dom";
 import {
@@ -7,122 +9,22 @@ import {
   CardHeader,
   CardTitle,
 } from "src/components/ui/card";
-import { ALBY_MIN_BALANCE, ALBY_SERVICE_FEE } from "src/constants";
-import { useAlbyBalance } from "src/hooks/useAlbyBalance";
-import { useAlbyMe } from "src/hooks/useAlbyMe";
-import { useApps } from "src/hooks/useApps";
-import { useChannels } from "src/hooks/useChannels";
-import { useInfo } from "src/hooks/useInfo";
-import { useNodeConnectionInfo } from "src/hooks/useNodeConnectionInfo";
-import { useTransactions } from "src/hooks/useTransactions";
+import { useOnboardingData } from "src/hooks/useOnboardingData";
 import { cn } from "src/lib/utils";
 
-function OnboardingChecklist() {
-  const { data: albyBalance } = useAlbyBalance();
-  const { data: albyMe } = useAlbyMe();
-  const { data: apps } = useApps();
-  const { data: channels } = useChannels();
-  const { data: info, hasChannelManagement, hasMnemonic } = useInfo();
-  const { data: nodeConnectionInfo } = useNodeConnectionInfo();
-  const { data: transactions } = useTransactions(false, 1);
+interface ChecklistItemProps {
+  title: string;
+  checked: boolean;
+  description: string;
+  to: string;
+}
 
-  const isLoading =
-    !albyMe ||
-    !apps ||
-    !channels ||
-    !info ||
-    !nodeConnectionInfo ||
-    !transactions ||
-    !albyBalance;
+function OnboardingChecklist() {
+  const { isLoading, checklistItems } = useOnboardingData();
 
   if (isLoading) {
-    return;
+    return null;
   }
-
-  const isLinked =
-    albyMe &&
-    nodeConnectionInfo &&
-    albyMe?.keysend_pubkey === nodeConnectionInfo?.pubkey;
-  const hasChannel =
-    !hasChannelManagement || (hasChannelManagement && channels.length > 0);
-  const hasBackedUp =
-    hasMnemonic &&
-    info &&
-    info.nextBackupReminder &&
-    new Date(info.nextBackupReminder).getTime() > new Date().getTime();
-  const hasCustomApp =
-    apps && apps.find((x) => x.name !== "getalby.com") !== undefined;
-  const hasTransaction = transactions.length > 0;
-
-  if (
-    isLinked &&
-    hasChannel &&
-    (!hasMnemonic || hasBackedUp) &&
-    hasCustomApp &&
-    hasTransaction
-  ) {
-    return;
-  }
-
-  const canMigrateAlbyFundsToNewChannel =
-    hasChannelManagement &&
-    info.backendType === "LDK" &&
-    albyBalance.sats * (1 - ALBY_SERVICE_FEE) >
-      ALBY_MIN_BALANCE + 50000; /* accomodate for onchain fees */
-
-  const checklistItems = [
-    {
-      title: "Open your first channel",
-      description:
-        "Establish a new Lightning channel to enable fast and low-fee Bitcoin transactions.",
-      checked: hasChannel,
-      to: canMigrateAlbyFundsToNewChannel
-        ? "/onboarding/lightning/migrate-alby"
-        : "/channels/outgoing",
-    },
-    {
-      title: "Send or receive your first payment",
-      description:
-        "Use your newly opened channel to make a transaction on the Lightning Network.",
-      checked: hasTransaction,
-      to: "/wallet",
-    },
-    {
-      title: "Link to your Alby Account",
-      description: "Link your lightning address & other apps to this Hub.",
-      checked: isLinked,
-      to: "/apps",
-    },
-    // TODO: enable when we can always migrate funds
-    /*{
-      title: "Migrate your balance to your Hub",
-      description: "Move your existing funds into self-custody.",
-      checked: !hasAlbyBalance,
-      to: "/onboarding/lightning/migrate-alby",
-    },*/
-    {
-      title: "Connect your first app",
-      description:
-        "Seamlessly connect apps and integrate your wallet with other apps from your Hub.",
-      checked: hasCustomApp,
-      to: "/appstore",
-    },
-    ...(hasMnemonic
-      ? [
-          {
-            title: "Backup your keys",
-            description:
-              "Secure your keys by creating a backup to ensure you don't lose access.",
-            checked: hasBackedUp,
-            to: "/settings/key-backup",
-          },
-        ]
-      : []),
-  ];
-
-  const sortedChecklistItems = checklistItems.sort(
-    (a, b) => (b && b.checked ? 1 : 0) - (a && a.checked ? 1 : 0)
-  );
 
   return (
     <Card>
@@ -134,12 +36,12 @@ function OnboardingChecklist() {
         </CardDescription>
       </CardHeader>
       <CardContent className="flex flex-col">
-        {sortedChecklistItems.map((item) => (
+        {checklistItems.map((item) => (
           <ChecklistItem
             key={item.title}
             title={item.title}
             description={item.description}
-            checked={!!item.checked}
+            checked={item.checked}
             to={item.to}
           />
         ))}
@@ -147,13 +49,6 @@ function OnboardingChecklist() {
     </Card>
   );
 }
-
-type ChecklistItemProps = {
-  title: string;
-  checked: boolean;
-  description: string;
-  to: string;
-};
 
 function ChecklistItem({
   title,
